add keyboard on main page
