+++ src/_scssVariables.scss
@@ -6,6 +6,8 @@ $main-color: #ffffff;
 $placeholder-color: rgba(255, 255, 255, 0.3);
 $action-color: #06f;
 $errorMessage-color: #ff0c46;
+$keyBlock-background: #1B1B1B;
+
 
 
 $authTitle-size: 23;
