--- src/components/DataBlock/DataInfo/_Datainfo.module.scss
@@ -1,10 +1,0 @@
-.dataInfo{
-    background-color: transparent;
-    background: #1B1B1B;
-    height: 2.625rem;
-    width: 17.5625rem;
-    border-radius: 11px;
-    color:  rgba(255, 255, 255, 0.3);
-    margin-bottom: 9px;
-    padding-left: 15px;
-}
