+++ src/components/KeyboardBlock/_KeyBoard.module.scss
@@ -1,0 +1,45 @@
+.KeyBoardBlock{
+    width: 820px;
+    display: flex;
+    flex-wrap: wrap;
+    color: #FFFFFF;
+    font-weight: 400;
+    margin-top: 5px;
+    gap: 7px;
+    div{
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        position: relative;
+    }
+    div:nth-child(54){
+        position: relative;
+        left: 21%;
+    }
+}
+
+.wrongKeyCode {
+	animation: borderRed .4s linear;
+}
+@keyframes borderRed {
+	0% {
+		border: 2px solid red;
+	}
+
+	100% {
+		border: none;
+	}
+}
+
+.keyCapPress {
+	animation: backgroundBlack .3s linear;
+}
+@keyframes backgroundBlack {
+	0% {
+		background: black;
+	}
+
+	100% {
+		background: #1B1B1B;
+	}
+}
