--- src/components/SimulatorBlock/KeyboardBlock/KeyBlock/KeyBlock.tsx
@@ -1,20 +1,0 @@
-import React, { FC } from 'react'
-import { IStylesOfKeyCaps } from '../../../../types/types'
-
-interface IKeyBlock {
-	styles: IStylesOfKeyCaps
-	elem: string
-	helpfulRecess?: IStylesOfKeyCaps
-	id: string
-}
-
-const KeyBlock: FC<IKeyBlock> = ({ styles, elem, helpfulRecess, id }) => {
-	return (
-		<div style={styles} id={id}>
-			{elem}
-			<div style={helpfulRecess}></div>
-		</div>
-	)
-}
-
-export const MemoizedKeyBlock = React.memo(KeyBlock)
