+++ src/components/SimulatorBlock/SimulatorBlock.tsx
@@ -1,15 +1,11 @@
 import React from "react";
-import { SimulatorStr } from "./SimulatorStrBlock/SimulatorStr";
-import { KeyBoardBlock } from './KeyboardBlock/KeyboardBlock';
-import { LevelBlock } from './LevelsBlock/LevelBlock';
+import { KeyBoardBlock } from '../KeyboardBlock/KeyboardBlock';
 import cl from "./_SimulatorBlock.module.scss";
 
 
 export const SimulatorBlock = () => {
   return (
     <div className={cl.MainContainer}>
-      <LevelBlock />
-      <SimulatorStr />
       <KeyBoardBlock />
     </div>
   );
