+++ src/pages/SimulatorPage/SimulatorPage.tsx
@@ -1,13 +1,11 @@
 import React from "react";
-import { Header } from '../../components/Header/Header';
 import { SimulatorBlock } from '../../components/SimulatorBlock/SimulatorBlock';
 
 export const SimulatorPage = () => {
   return (
-		<div className='App'>
-      <Header />
+		<>
       <SimulatorBlock />
-    </div>
+    </>
   );
 };
